pyproject.toml
[tool.poetry]
name = "dndnd"
version = "0.1.0"
description = ""
authors = ["Jeremy Fisher <jeremyadamsfisher@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.87"
openai = "^0.26.5"
streamlit-chat = "^0.0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
